[ignore] Refactor adaptive serializer to support indentation. Correctly output atomic values if no other rules apply.
